windows/requirements.txt
pyserial>=3.5
pynput>=1.7.6
mss>=9.0
Pillow>=10.0
